#73 - separate table definitions from daos
