+++ app/daos/AccessRightDao.scala
@@ -1,7 +1,9 @@
 package daos
 
-import java.net.{URI, URISyntaxException}
+import java.util.UUID
 import javax.inject.Inject
+
+import daos.mariaDB.{AccessRightTableDef, TaskAccessRightTableDef, TaskTableDef, UserTaskTableDef}
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
@@ -9,8 +11,7 @@ import play.api.Play
 import play.api.db.slick.DatabaseConfigProvider
 import slick.driver.JdbcProfile
 import slick.driver.MySQLDriver.api._
-import models.{AccessRight, HttpMethod}
-import models.HttpMethod.HttpMethod
+import models.AccessRight
 
 /**
   * Created bei jottmann on 29.08.2017
@@ -25,44 +26,14 @@ trait AccessRightDao {
   def delete(id:Long): Future[Int]
 }
 
-
-
-class AccessRightTableDef(tag: Tag) extends Table[AccessRight](tag, "AccessRight") {
-  def id = column[Long]("id", O.PrimaryKey,O.AutoInc)
-  def uri = column[URI]("uri")
-  def method = column[HttpMethod]("method")
-  def name = column[String]("name")
-  def description = column[String]("description")
-  def service = column[String]("service")
-
-  implicit val uriColumnType = MappedColumnType.base[URI, String](
-    { uri => uri.toASCIIString },
-    { str =>
-      try {
-        new URI(str)
-      } catch {
-        case e:URISyntaxException => throw new SlickException(s"Invalid URI value: $str", e)
-      }
-    }
-  )
-
-  implicit val httpMethodColumnType = MappedColumnType.base[HttpMethod, String](
-    {httpMethod => httpMethod.toString},
-    {
-      case "POST" => HttpMethod.POST
-      case "GET" => HttpMethod.GET
-    }
-  )
-
-  def * =
-    (id, uri, method, name.?, description.?, service) <>((AccessRight.mapperTo _).tupled, AccessRight.unapply)
-}
-
 class MariadbAccessRightDao @Inject()(dbConfigProvider: DatabaseConfigProvider) extends AccessRightDao {
   val dbConfig = DatabaseConfigProvider.get[JdbcProfile](Play.current)
 
   val accessRights = TableQuery[AccessRightTableDef]
   val taskAccessRights = TableQuery[TaskAccessRightTableDef]
+  val tasks = TableQuery[TaskTableDef]
+  val userTasks = TableQuery[UserTaskTableDef]
+
 
   def all(): Future[Seq[AccessRight]] = dbConfig.db.run(accessRights.result)
